Bugfix: read scatter single test multiple ranks
Change kvtree_read_scatter_single_test to initialize kvtree to NULL
to allow ranks that never create kvtree to call `kvtree_delete()`.

It seems on some platforms `&kvtree` shows as an error rather than
NULL when it was never created, allowing the checks in
`kvtree_delete()` to pass and a delete is still attempted, resulting
in a segfault.